+++ calva/nrepl/jack-in.ts
@@ -1,34 +1,16 @@
 import * as vscode from "vscode";
 import * as utilities from "../utilities";
 import * as fs from "fs";
+import * as path from "path";
 import * as state from "../state"
-import connector from "../connector";
+import * as connector from "../connector";
 import statusbar from "../statusbar";
-import * as shadow from "../shadow"
-//const { parseEdn, parseForms } = require('../../cljs-out/cljs-lib');
 import { parseEdn, parseForms } from "../../cljs-out/cljs-lib";
 
 const isWin = /^win/.test(process.platform);
 
-/** Finds a file in PATH */
-function findInPath(name: string) {
-    const paths = process.env.PATH.split(isWin ? ";" : ":");
-    for (let path of paths) {
-        let fullPath = path + (isWin ? "\\" : "/") + name;
-        if (fs.existsSync(fullPath))
-            return fullPath;
-    }
-}
-
-/** If this looks like a string and we are under windows, escape it in a powershell compatible way. */
-function escapeString(str: string) {
-    if (str.startsWith('"') && str.endsWith('"') && isWin)
-        return str.replace(/\\"/g, "\\`\"");
-    return str;
-}
-
-export function detectProjectType(): string[] {
-    let rootDir = utilities.getProjectDir(),
+export async function detectProjectType(): Promise<string[]> {
+    let rootDir = connector.getProjectRoot(),
         cljProjTypes = []
     for (let clj in projectTypes) {
         try {
@@ -59,32 +41,29 @@ const leinDependencies = {
 const middleware = ["cider.nrepl/cider-middleware"];
 const cljsMiddleware = ["cider.piggieback/wrap-cljs-repl"];
 
-const initEval = '"(require (quote cider-nrepl.main)) (cider-nrepl.main/init [\\"cider.nrepl/cider-middleware\\", \\"cider.piggieback/wrap-cljs-repl\\"])"';
-
-const projectTypes: {
-    [id: string]: {
-        name: string,
-        cljsTypes: string[],
-        cmd: string,
-        winCmd: string,
-        commandLine: (includeCljs: boolean) => any,
-        useWhenExists: string,
-        useShell?: boolean
-    }
-} = {
+const projectTypes: { [id: string]: connector.ProjectType } = {
     "lein": {
         name: "Leiningen",
         cljsTypes: ["Figwheel", "Figwheel Main"],
         cmd: "lein",
-        winCmd: "lein.bat",
-        useShell: true,
+        winCmd: "cmd.exe",
         useWhenExists: "project.clj",
+        nReplPortFile: () => {
+            return connector.nreplPortFile(".nrepl-port");
+        },
+        /** Build the Commandline args for a lein-project. 
+         * 1. Parsing the project.clj
+         * 2. Let the user choose a alias
+         * 3. Let the user choose profiles to use
+         * 4. Add nedded middleware deps to args
+         * 5. Add all profiles choosed by the user
+         * 6. Use alias if selected otherwise repl :headless
+        */
         commandLine: async (includeCljs) => {
             let out: string[] = [];
             let dependencies = includeCljs ? { ...leinDependencies, ...figwheelDependencies } : leinDependencies;
             let keys = Object.keys(dependencies);
-
-            let data = fs.readFileSync(utilities.getProjectDir() + "/project.clj", 'utf8').toString();
+            let data = fs.readFileSync(path.resolve(connector.getProjectRoot(), "project.clj"), 'utf8').toString();
             let parsed;
             try {
                 parsed = parseForms(data);
@@ -93,15 +72,20 @@ const projectTypes: {
                 throw e;
             }
             let profiles: string[] = [];
+            let alias: string;
             const defproject = parsed.find(x => x[0] == "defproject");
-            if (defproject != undefined) {
-                let profilesIndex = defproject.indexOf("profiles");
-                if (profilesIndex > -1) {
+
+            if (defproject) {
+                let aliasesIndex = defproject.indexOf("aliases");
+                if (aliasesIndex > -1) {
                     try {
-                        const profilesMap = defproject[profilesIndex + 1];
-                        profiles = [...profiles, ...Object.keys(profilesMap).map((v, k) => { return ":" + v })];
-                        if (profiles.length) {
-                            profiles = await utilities.quickPickMulti({ values: profiles, saveAs: "lein-cli-profiles", placeHolder: "Pick any profiles to launch with" });
+                        let aliases: string[] = [];
+                        const aliasesMap = defproject[aliasesIndex + 1];
+                        aliases = [...profiles, ...Object.keys(aliasesMap).map((v, k) => { return v })];
+                        if (aliases.length) {
+                            aliases.unshift("No alias");
+                            alias = await utilities.quickPickSingle({ values: aliases, saveAs: `${connector.getProjectRoot()}/lein-cli-alias`, placeHolder: "Choose alias to run" });
+                            alias = (alias == "No alias") ? undefined : alias;
                         }
                     } catch (error) {
                         vscode.window.showErrorMessage("The project.clj file is not sane. " + error.message);
@@ -110,27 +94,55 @@ const projectTypes: {
                 }
             }
 
+            if (defproject != undefined) {
+                let profilesIndex = defproject.indexOf("profiles");
+                if (profilesIndex > -1) {
+                    try {
+                        const profilesMap = defproject[profilesIndex + 1];
+                        profiles = [...profiles, ...Object.keys(profilesMap).map((v, k) => { return ":" + v })];
+                        if (profiles.length) {
+                            profiles = await utilities.quickPickMulti({ values: profiles, saveAs: `${connector.getProjectRoot()}/lein-cli-profiles`, placeHolder: "Pick any profiles to launch with" });
+                        }
+                    } catch (error) {
+                        vscode.window.showErrorMessage("The project.clj file is not sane. " + error.message);
+                        console.log(error);
+                    }
+                }
+            }
+
+            if (isWin) {
+                out.push("/d", "/c", "lein");
+            }
+            const q = isWin ? '' : "'",
+            dQ = '"',
+            s = isWin ? "^ " : " ";
+
             for (let i = 0; i < keys.length; i++) {
                 let dep = keys[i];
-                out.push("update-in", ":dependencies", "conj", `"[${dep} \\"${dependencies[dep]}\\"]"`, '--');
+                out.push("update-in", ":dependencies", "conj", `${q + "[" + dep + dQ + dependencies[dep] + dQ + "]" + q}`, '--');
             }
 
             keys = Object.keys(leinPluginDependencies);
             for (let i = 0; i < keys.length; i++) {
                 let dep = keys[i];
-                out.push("update-in", ":plugins", "conj", `"[${dep} \\"${leinPluginDependencies[dep]}\\"]"`, '--');
+                out.push("update-in", ":plugins", "conj", `${q + "[" + dep + dQ + leinPluginDependencies[dep] + dQ + "]" + q}`, '--');
             }
 
             const useMiddleware = includeCljs ? [...middleware, ...cljsMiddleware] : middleware;
             for (let mw of useMiddleware) {
-                out.push("update-in", '"[:repl-options :nrepl-middleware]"', "conj", `"[\\"${mw.replace('"', '\\"')}\\"]"`, '--');
+                out.push("update-in", `${q + '[:repl-options' + s + ':nrepl-middleware]' + q}`, "conj", `'["${mw}"]'`, '--');
             }
 
             if (profiles.length) {
                 out.push("with-profile", profiles.map(x => `+${x.substr(1)}`).join(','));
             }
-            //out.push("update-in", ":middleware", "conj", `cider-nrepl.plugin/middleware`, '--')
-            out.push("repl", ":headless");
+
+            if (alias) {
+                out.push(alias);
+            } else {
+                out.push("repl", ":headless");
+            }
+            
             return out;
         }
     },
@@ -139,7 +151,6 @@ const projectTypes: {
         name: "Boot",
         cmd: "boot",
         winCmd: "boot.exe",
-        useShell: true,
         useWhenExists: "build.boot",      
         commandLine: () => {
             let out: string[] = [];
@@ -153,11 +164,22 @@ const projectTypes: {
         name: "Clojure CLI",
         cljsTypes: ["Figwheel", "Figwheel Main"],
         cmd: "clojure",
-        winCmd: "clojure.ps1",
+        winCmd: "powershell.exe",
         useWhenExists: "deps.edn",
+        nReplPortFile: () => {
+            return connector.nreplPortFile(".nrepl-port");
+        },
+        /** Build the Commandline args for a clj-project.
+         * 1. Read the deps.edn and parsed it 
+         * 2. Present the user all found aliases
+         * 3. Define needed dependencies and middlewares used by calva
+         * 4. Check if the selected aliases have main-opts
+         * 5. If main-opts in alias => just use aliases
+         * 6. if no main-opts => supply our own main to run nrepl with middlewares
+         */
         commandLine: async (includeCljs) => {
             let out: string[] = [];
-            let data = fs.readFileSync(utilities.getProjectDir() + "/deps.edn", 'utf8').toString();
+            let data = fs.readFileSync(path.join(connector.getProjectRoot(), "deps.edn"), 'utf8').toString();
             let parsed;
             try {
                 parsed = parseEdn(data);
@@ -165,17 +187,39 @@ const projectTypes: {
                 vscode.window.showErrorMessage("Could not parse deps.edn");
                 throw e;
             }
-            let aliases = [];
+            let aliases:string[] = [];
             if (parsed.aliases != undefined) {
-                aliases = await utilities.quickPickMulti({ values: Object.keys(parsed.aliases).map(x => ":" + x), saveAs: "clj-cli-aliases", placeHolder: "Pick any aliases to launch with" });
+                aliases = await utilities.quickPickMulti({ values: Object.keys(parsed.aliases).map(x => ":" + x), saveAs: `${connector.getProjectRoot()}/clj-cli-aliases`, placeHolder: "Pick any aliases to launch with" });
             }
 
             const dependencies = includeCljs ? { ...cliDependencies, ...figwheelDependencies } : cliDependencies,
                 useMiddleware = includeCljs ? [...middleware, ...cljsMiddleware] : middleware;
-            const aliasesOption = aliases.length > 0 ? `-A${aliases.join("")}` : ''
+            const aliasesOption = aliases.length > 0 ? `-A${aliases.join("")}` : '';
+            let aliasHasMain:boolean = false;
+            for (let ali in aliases) {
+                let aliasKey = aliases[ali].substr(1);
+                let alias =  parsed.aliases[aliasKey];
+                aliasHasMain = (alias["main-opts"] != undefined);
+                if (aliasHasMain)
+                    break;
+            }
+
+            const dQ = isWin ? '""' : '"';
             for (let dep in dependencies)
-                out.push(dep + " {:mvn/version \\\"" + dependencies[dep] + "\\\"}")
-            return ["-Sdeps", `"${"{:deps {" + out.join(' ') + "}}"}"`, aliasesOption, "-m", "nrepl.cmdline", "--middleware", `"[${useMiddleware.join(' ')}]"`]
+                out.push(dep + ` {:mvn/version ${dQ}${dependencies[dep]}${dQ}}`)
+
+            let args = ["-Sdeps", `'${"{:deps {" + out.join(' ') + "}}"}'`];
+            
+            if (aliasHasMain) {
+                args.push(aliasesOption);
+            } else {
+                args.push(aliasesOption, "-m", "nrepl.cmdline", "--middleware", `"[${useMiddleware.join(' ')}]"`);
+            }
+
+            if (isWin) {
+                args.unshift("clojure");
+            }
+            return args;
         }
     },
     "shadow-cljs": {
@@ -183,14 +227,20 @@ const projectTypes: {
         cljsTypes: [],
         cmd: "npx",
         winCmd: "npx.cmd",
-        useShell: true,
         useWhenExists: "shadow-cljs.edn",
+        nReplPortFile: () => {
+            return connector.nreplPortFile(path.join(".shadow-cljs", "nrepl.port"));
+        },
+        /**
+         *  Build the Commandline args for a shadow-project.
+         */
         commandLine: async (_includeCljs) => {
             let args: string[] = [];
             for (let dep in shadowDependencies)
                 args.push("-d", dep + ":" + shadowDependencies[dep]);
 
-            let builds = await utilities.quickPickMulti({ values: shadow.shadowBuilds().filter(x => x[0] == ":"), placeHolder: "Select builds to start", saveAs: "shadowcljs-jack-in" })
+            const shadowBuilds = await connector.shadowBuilds();
+            let builds = await utilities.quickPickMulti({ values: shadowBuilds.filter(x => x[0] == ":"), placeHolder: "Select builds to start", saveAs: `${connector.getProjectRoot()}/shadowcljs-jack-in` })
             if (!builds || !builds.length)
                 return;
             return ["shadow-cljs", ...args, "watch", ...builds];
@@ -208,33 +258,73 @@ function getProjectTypeForName(name: string) {
 let watcher: fs.FSWatcher;
 const TASK_NAME = "Calva Jack-in";
 
-vscode.tasks.onDidStartTaskProcess(e => {
-    if (e.execution.task.name == TASK_NAME) {
+async function executeJackInTask(projectType: connector.ProjectType, projectTypeSelection: any, executable: string, args: any, cljTypes: string[], outputChannel: vscode.OutputChannel) {
+    state.cursor.set("launching", projectTypeSelection);
+    statusbar.update();
+    const nReplPortFile = projectType.nReplPortFile();
+    const env = { ...process.env, ...state.config().jackInEnv } as {
+        [key: string]: string;
+    };
+    const execution = isWin ?
+        new vscode.ProcessExecution(executable, args, {
+            cwd: connector.getProjectRoot(),
+            env: env,
+        }) :
+        new vscode.ShellExecution(executable, args, {
+            cwd: connector.getProjectRoot(),
+            env: env,
+        });
+    const taskDefinition: vscode.TaskDefinition = {
+        type: isWin ? "process" : "shell",
+        label: "Calva: Jack-in"
+    };
+    const task = new vscode.Task(taskDefinition, connector.getProjectWsFolder(), TASK_NAME, "Calva", execution);
+
+    state.analytics().logEvent("REPL", "JackInExecuting", JSON.stringify(cljTypes)).send();
+
+    vscode.tasks.executeTask(task).then((v) => {
+        // Create a watcher to wait for the nREPL port file to appear with new content, and connect + open the repl window at that point.
+        const portFileDir = path.dirname(nReplPortFile),
+            portFileBase = path.basename(nReplPortFile);
         if (watcher != undefined) {
             watcher.removeAllListeners();
         }
-        // Create a watcher to wait for the nREPL port file to appear, and connect + open the repl window at that point.
-        watcher = fs.watch(utilities.getProjectDir(), async (eventType, filename) => {
-            if (filename == ".nrepl-port") {
+      
+        watcher = fs.watch(portFileDir, async (eventType, fileName) => {
+            if (fileName == portFileBase) {
+                if (!fs.existsSync(nReplPortFile)) {
+                    return;
+                }
+                const port = fs.readFileSync(nReplPortFile, 'utf8');
+                if (!port) { // On Windows we get two events, one for file creation and one for the change of content
+                    return;  // If there is no port to be read yet, wait for the next event instead.
+                }
                 const chan = state.outputChannel();
                 setTimeout(() => { chan.show() }, 1000);
                 state.cursor.set("launching", null);
-                watcher.close();
+                watcher.removeAllListeners();
                 await connector.connect(true, true);
                 chan.appendLine("Jack-in done.\nUse the VS Code task management UI to control the life cycle of the Jack-in task.");
             }
-        })
-    }
-});
+        });
+    }, (reason) => {
+        watcher.removeAllListeners();
+        outputChannel.appendLine("Error in Jack-in: " + reason);
+    });
+}
 
 export async function calvaJackIn() {
     const outputChannel = state.outputChannel();
-
+    try {
+        await connector.initProjectDir();
+    } catch {
+        return;
+    }
     state.analytics().logEvent("REPL", "JackInInitiated").send();
     outputChannel.appendLine("Jacking in...");
 
     // figure out what possible kinds of project we're in
-    let cljTypes = detectProjectType();
+    let cljTypes = await detectProjectType();
     if (cljTypes.length == 0) {
         vscode.window.showErrorMessage("Cannot find project, no project.clj, deps.edn or shadow-cljs.edn.");
         state.analytics().logEvent("REPL", "JackInInterrupted", "FailedFindingProjectType").send();
@@ -254,12 +344,11 @@ export async function calvaJackIn() {
             menu.push(`${projectTypes[clj].name} + ${cljs}`);
         }
     }
-    let projectTypeSelection = await utilities.quickPickSingle({ values: menu, placeHolder: "Please select a project type", saveAs: "jack-in-type", autoSelect: true });
+    let projectTypeSelection = await utilities.quickPickSingle({ values: menu, placeHolder: "Please select a project type", saveAs: `${connector.getProjectRoot()}/jack-in-type`, autoSelect: true });
     if (!projectTypeSelection) {
         state.analytics().logEvent("REPL", "JackInInterrupted", "NoProjectTypePicked").send();
         return;
     }
-
 
     // Resolve the selection to an entry in projectTypes
     const projectTypeName: string = projectTypeSelection.replace(/ \+ .*$/, "");
@@ -273,92 +362,10 @@ export async function calvaJackIn() {
         return;
     }
 
-    // Now look in our $PATH variable to check the appropriate command exists.
-    const cmd = isWin ? projectType.winCmd : projectType.cmd;
-    let executable = findInPath(cmd);
-    let integrated_shell;
-
-    if (!executable) {
-        // It doesn't, do not proceed
-        state.analytics().logEvent("REPL", "JackInInterrupted", "CommandNotInPath").send();
-        vscode.window.showErrorMessage(cmd + " is not on your PATH, please add it.")
-        return;
-    }
+    let executable = isWin ? projectType.winCmd : projectType.cmd;
 
     // Ask the project type to build up the command line. This may prompt for further information.
     let args = await projectType.commandLine(selectedCljsType != "");
 
-    let shellSettingsShouldBeChangedByUs = false;
-    let shellSettingsChangedByUs = false;
-
-    if (executable.endsWith(".ps1")) {
-        // launch powershell scripts through powershell, doing crazy powershell escaping.
-        args = args.map(escapeString) as Array<string>;
-        args.unshift(executable);
-        executable = "powershell.exe";
-    } else if (executable.endsWith(".bat")) {
-        // Current workaround for the not working powershell etc. changes to cmd.exe and later back to whaterver was set before
-        let windowsterminalssettings = vscode.workspace.getConfiguration("terminal.integrated.shell").inspect("windows");
-        integrated_shell = windowsterminalssettings.workspaceFolderValue || windowsterminalssettings.workspaceValue;
-        if (!integrated_shell || !integrated_shell.endsWith("cmd.exe")) {
-            shellSettingsShouldBeChangedByUs = true;
-            outputChannel.appendLine("Jack-in needs to use cmd.exe to work. Allow the (temporary) switch, please.");
-            await vscode.workspace.getConfiguration()
-                .update("terminal.integrated.shell.windows", "C:\\Windows\\System32\\cmd.exe")
-                .then(
-                    () => {
-                        shellSettingsChangedByUs = true;
-                    },
-                    (reason) => {
-                        outputChannel.appendLine("Jack-in aborted, since it won't work without using cmd.exe");
-                    }
-                );
-        }
-    }
-
-    if (shellSettingsShouldBeChangedByUs && !shellSettingsChangedByUs) {
-        state.analytics().logEvent("REPL", "JackInInterrupted", "Powershell user wont allow change of shell setting").send();
-        return;
-    }
-    state.cursor.set("launching", projectTypeSelection)
-    statusbar.update();
-
-    const nreplPortFile = connector.nreplPortFile();
-    if (nreplPortFile && fs.existsSync(nreplPortFile)) {
-        fs.unlinkSync(nreplPortFile);
-    }
-
-    const env = { ...process.env, ...state.config().jackInEnv } as { [key: string]: string; };
-
-    const execution = new vscode.ShellExecution(executable, args, {
-        cwd: utilities.getProjectDir(),
-        env: env
-    });
-
-    const taskDefinition: vscode.TaskDefinition = {
-        type: "shell",
-        label: "Calva: Jack-in"
-    };
-
-    const folder = vscode.workspace.workspaceFolders[0];
-    const task = new vscode.Task(taskDefinition, folder, TASK_NAME, "Calva", execution);
-
-    state.analytics().logEvent("REPL", "JackInExecuting", JSON.stringify(cljTypes)).send();
-    vscode.tasks.executeTask(task).then(
-        (v) => {
-        },
-        (reason) => {
-            watcher.close()
-            outputChannel.appendLine("Error in Jack-in: " + reason);
-        });
-
-    if (executable.endsWith(".bat") && shellSettingsChangedByUs) {
-        setTimeout(() => { // Need to give the task shell time to start
-            vscode.workspace.getConfiguration()
-                .update("terminal.integrated.shell.windows", integrated_shell)
-                .then(() => {
-                    integrated_shell = undefined;
-                });
-        }, 2000);
-    }
+    executeJackInTask(projectType, projectTypeSelection, executable, args, cljTypes, outputChannel);
 }